Add StateOperator
This is the "main" operator for a Keeper. It allows a programmer to
define how a state should be saved and how to serve clients' queries.

The basic idea is that StateOperator can take in two streams (one with
updates to the state and one with queries from clients). For each of the
streams user can then define the operator implementation in almost the
same way as it would be done with regular Timely operators but also
having access to some shared state. Output streams can be later retrieved
from StateOperator object. To create a StateOperator instance one needs
to use StateOperatorBuilder.

Test in src/core/state.rs introduced in this commit shows an example use of
StateOperator.