Remove Query and ResponseTuple
They were the structs that client and Keeper would exchange, but it
doesn't really make sense to have a specialized struct for this, if it
has only one field inside it. Instead each Keeper is now parametrized by
two types: Query and Request type. When a client wants to query a Keeper
it needs to use the same types as the ones that the Keeper is
parametrized with.