Added KeeperQuery and KeeperResponse
Those are the types that will be sent between a Keeper an its clients.
They support the new "state, update, point query" idea where each Keeper
always exposes the interface to get all its current state and/or
subscribe to updates to the state. Such Keeper can also allow for point
queries on the state it holds.

Since client side is not updated yet, the current implementations don't
work with this commit.