Update to the "new Keeper concept"
The new Keeper concept simplifies it a bit - in the previous version the
idea was for Keeper to allow for implementing standing queries on its
state.
Current version allows only for point/adhoc queries but additionaly it
support "StateRequests". With state request client can subscribe to
updates to the state and/or get the current state of the Keeper.
This commit modifies Keeper's state to support this concept and updates
implementations to use the new interface (though none of them yet
presents the posibilities of "Update" request).