Add `strymon_job` crate and new publisher logic
This implements a Timely stream publisher which keeps track of both
the "upper" and "lower" frontier of the published stream, allowing
subscribers to filter out incomplete epochs.

In addition, this publisher now contains all its logic in a separate
thread, so it can be scheduled even when the Timely computation is
idle.

Signed-off-by: Sebastian Wicki <swicki@inf.ethz.ch>